try to add draggable but cannot update the value yet
