+++ package.json
@@ -12,6 +12,7 @@
     "firebase": "^6.2.3",
     "vue": "^2.6.10",
     "vue-router": "^3.0.3",
+    "vuedraggable": "^2.23.0",
     "vuetify": "^1.5.5"
   },
   "devDependencies": {
